voice-module/requirements.txt
# Voice Interaction Backend Dependencies
pipecat-ai[daily,openai]>=0.0.40
daily-python>=0.10.0
redis[hiredis]>=5.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
loguru>=0.7.2
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
httpx>=0.26.0
numpy>=1.24.0
google-genai>=0.5.0
browserbase>=1.0.0
playwright>=1.40.0
